randr12: use clocks in kHz throughout

+++ src/nv_proto.h
@@ -85,8 +85,8 @@ void NVPointerMoved(int index, int x, int y);
 /* in nv_bios.c */
 unsigned int NVParseBios(ScrnInfoPtr pScrn);
 void link_head_and_output(ScrnInfoPtr pScrn, int head, int dcb_entry, Bool overrideval);
-void call_lvds_script(ScrnInfoPtr pScrn, int head, int dcb_entry, enum LVDS_script script, uint16_t pxclk);
-void run_tmds_table(ScrnInfoPtr pScrn, int dcb_entry, int head, uint16_t pxclk);
+void call_lvds_script(ScrnInfoPtr pScrn, int head, int dcb_entry, enum LVDS_script script, int pxclk);
+void run_tmds_table(ScrnInfoPtr pScrn, int dcb_entry, int head, int pxclk);
 Bool get_pll_limits(ScrnInfoPtr pScrn, enum pll_types plltype, struct pll_lims *pll_lim);
 uint32_t getMNP_single(ScrnInfoPtr pScrn, uint32_t clk, int *NM, int *log2P);
 uint32_t getMNP_double(ScrnInfoPtr pScrn, struct pll_lims *pll_lim, uint32_t clk, int *NM1, int *NM2, int *log2P);
